requirements.txt
ta
python-binance
numpy
pandas
time
json
sklearn
xlsxwriter
